+++ HLTrigger/Configuration/python/HLT_75e33/modules/hltHgcalSoARecHitsLayerClustersProducerHSci_cfi.py
@@ -7,9 +7,9 @@ hltHgcalSoARecHitsLayerClustersProducerHSci = cms.EDProducer("HGCalCLUEsteringLa
     ),
     hgcalRecHitsSoA = cms.InputTag("hltHgcalSoARecHitsProducerHSci"),
     detector = cms.string('BH'),
-    deltac = cms.double(0.0315),
-    kappa = cms.double(9),
-    outlierDeltaFactor = cms.double(2.0)
+    deltac = cms.float(0.0315),
+    kappa = cms.float(9),
+    outlierDeltaFactor = cms.float(2.0)
 )
 
 hltHgcalSoARecHitsLayerClustersProducerHSciSerialSync = makeSerialClone(hltHgcalSoARecHitsLayerClustersProducerHSci,
