+++ HLTrigger/Configuration/python/HLT_75e33/modules/hltHgcalSoARecHitsLayerClustersProducerHSi_cfi.py
@@ -7,9 +7,9 @@ hltHgcalSoARecHitsLayerClustersProducerHSi = cms.EDProducer("HGCalCLUEsteringLay
     ),
     hgcalRecHitsSoA = cms.InputTag("hltHgcalSoARecHitsProducerHSi"),
     detector = cms.string('FH'),
-    deltac = cms.double(1.3),
-    kappa = cms.double(9),
-    outlierDeltaFactor = cms.double(2.0)
+    deltac = cms.float(1.3),
+    kappa = cms.float(9),
+    outlierDeltaFactor = cms.float(2.0)
 )
 
 hltHgcalSoARecHitsLayerClustersProducerHSiSerialSync = makeSerialClone(hltHgcalSoARecHitsLayerClustersProducerHSi,
